// src/app/api/models/project-address.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProjectAddress {
  city?: string | null;
  country?: string | null;
  houseNumber?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  postalCode?: string | null;
  street?: string | null;
}